mergeParts

+++ BookStore/BookStore/Service/Category/CategoryService.cs
@@ -23,11 +23,32 @@ namespace BookStore.Service.Category
             return await _context.Categories.ToListAsync();
         }
 
+        public async Task<Models.Category> GetCategoryById(int id)
+        {
+            return await _context.Categories.FindAsync(id);
+        }
+
+        public async Task EditCategory(Models.Category updatedCategory)
+        {
+            var category = await _context.Categories.FindAsync(updatedCategory.Id);
+            if (category != null)
+            {
+                category.Name = updatedCategory.Name;
+                await _context.SaveChangesAsync();
+            } 
+        }
+
         public async Task DeleteCategory(int id)
         {
             var category = await _context.Categories.FindAsync(id);
-            _context.Categories.Remove(category);
-            await _context.SaveChangesAsync();
+
+            if(category != null)
+            {
+                var relatedBookCategories = _context.BookCategories.Where(bc => bc.CategoryId == id);
+                _context.BookCategories.RemoveRange(relatedBookCategories);
+                _context.Categories.Remove(category);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
